+++ crates/viewer/re_renderer_examples/depth_cloud.rs
@@ -13,7 +13,7 @@
 //! cargo run-wasm --example depth_cloud
 //! ```
 
-// TODO(#3408): remove unwrap()
+// TODO(#6330): remove unwrap()
 #![allow(clippy::unwrap_used)]
 
 use std::f32::consts::TAU;