+++ crates/viewer/re_renderer_examples/framework.rs
@@ -1,6 +1,6 @@
 //! Example framework
 
-// TODO(#3408): remove unwrap()
+// TODO(#6330): remove unwrap()
 #![allow(clippy::unwrap_used)]
 
 use std::sync::Arc;