+++ rerun_cpp/src/rerun/blueprint/components/latest_at_queries.hpp
@@ -18,7 +18,7 @@ namespace arrow {
 } // namespace arrow
 
 namespace rerun::blueprint::components {
-    /// **Component**: Configuration for latest at queries.
+    /// **Component**: Configuration for latest-at queries.
     ///
     /// Note: configuration as saved on a per-timeline basis.
     struct LatestAtQueries {